Move net and threading out
