+++ c-scape/src/lib.rs
@@ -38,8 +38,6 @@ mod unwind;
 use alloc::borrow::ToOwned;
 use alloc::boxed::Box;
 use alloc::format;
-#[cfg(feature = "sync-resolve")]
-use alloc::string::ToString;
 use core::convert::TryInto;
 use core::ffi::c_void;
 #[cfg(not(target_arch = "riscv64"))]
@@ -48,37 +46,14 @@ use core::mem::align_of;
 use core::mem::{size_of, zeroed};
 use core::ptr::{self, null, null_mut};
 use core::slice;
-#[cfg(feature = "threads")]
-use core::sync::atomic::Ordering::SeqCst;
-#[cfg(feature = "threads")]
-use core::sync::atomic::{AtomicBool, AtomicU32};
 use errno::{set_errno, Errno};
 use error_str::error_str;
-use libc::{c_char, c_int, c_long, c_uint, c_ulong};
-#[cfg(feature = "threads")]
-use parking_lot::lock_api::{RawMutex as _, RawRwLock};
-#[cfg(feature = "threads")]
-use parking_lot::RawMutex;
-use rustix::fd::{BorrowedFd, IntoRawFd};
+use libc::{c_char, c_int, c_long, c_ulong};
 use rustix::ffi::ZStr;
 use rustix::fs::{AtFlags, Mode, OFlags};
-#[cfg(feature = "net")]
-use rustix::net::{
-    AcceptFlags, AddressFamily, Ipv4Addr, Ipv6Addr, Protocol, RecvFlags, SendFlags, Shutdown,
-    SocketAddrAny, SocketAddrStorage, SocketFlags, SocketType,
-};
-#[cfg(feature = "sync-resolve")]
-use {
-    rustix::net::{IpAddr, SocketAddrV4, SocketAddrV6},
-    sync_resolve::resolve_host,
-};
 
 // fs
 mod fs;
-
-// process
-#[cfg(not(target_os = "wasi"))]
-mod process;
 
 // io
 mod io;
@@ -86,6 +61,18 @@ mod io;
 // mmap
 #[cfg(not(target_os = "wasi"))]
 mod mmap;
+
+// net
+#[cfg(feature = "net")]
+mod net;
+
+// process
+#[cfg(not(target_os = "wasi"))]
+mod process;
+
+// threads
+#[cfg(feature = "threads")]
+mod threads;
 
 // errno
 
@@ -110,766 +97,6 @@ unsafe extern "C" fn __xpg_strerror_r(errnum: c_int, buf: *mut c_char, buflen: u
     let out = slice::from_raw_parts_mut(buf.cast::<u8>(), min);
     out.copy_from_slice(message.as_bytes());
     out[out.len() - 1] = b'\0';
-    0
-}
-
-// net
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn accept(
-    fd: c_int,
-    addr: *mut SocketAddrStorage,
-    len: *mut libc::socklen_t,
-) -> c_int {
-    // FIXME(#95) layout of addr doesn't match signature on i686
-    // uncomment once it does:
-    // libc!(libc::accept(fd, checked_cast!(addr), len));
-    libc!(libc::accept(fd, addr.cast(), len));
-
-    match convert_res(rustix::net::acceptfrom(&BorrowedFd::borrow_raw(fd))) {
-        Some((accepted_fd, from)) => {
-            if let Some(from) = from {
-                let encoded_len = from.write(addr);
-                *len = encoded_len.try_into().unwrap();
-            }
-            accepted_fd.into_raw_fd()
-        }
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn accept4(
-    fd: c_int,
-    addr: *mut SocketAddrStorage,
-    len: *mut libc::socklen_t,
-    flags: c_int,
-) -> c_int {
-    // FIXME(#95) layout of addr doesn't match signature on i686
-    // uncomment once it does:
-    // libc!(libc::accept4(fd, checked_cast!(addr), len, flags));
-    libc!(libc::accept4(fd, addr.cast(), len, flags));
-
-    let flags = AcceptFlags::from_bits(flags as _).unwrap();
-    match convert_res(rustix::net::acceptfrom_with(
-        &BorrowedFd::borrow_raw(fd),
-        flags,
-    )) {
-        Some((accepted_fd, from)) => {
-            if let Some(from) = from {
-                let encoded_len = from.write(addr);
-                *len = encoded_len.try_into().unwrap();
-            }
-            accepted_fd.into_raw_fd()
-        }
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn bind(
-    sockfd: c_int,
-    addr: *const SocketAddrStorage,
-    len: libc::socklen_t,
-) -> c_int {
-    // FIXME(#95) layout of addr doesn't match signature on i686
-    // uncomment once it does:
-    // libc!(libc::bind(sockfd, checked_cast!(addr), len));
-    libc!(libc::bind(sockfd, addr.cast(), len));
-
-    let addr = match convert_res(SocketAddrAny::read(addr, len.try_into().unwrap())) {
-        Some(addr) => addr,
-        None => return -1,
-    };
-    match convert_res(match addr {
-        SocketAddrAny::V4(v4) => rustix::net::bind_v4(&BorrowedFd::borrow_raw(sockfd), &v4),
-        SocketAddrAny::V6(v6) => rustix::net::bind_v6(&BorrowedFd::borrow_raw(sockfd), &v6),
-        SocketAddrAny::Unix(unix) => rustix::net::bind_unix(&BorrowedFd::borrow_raw(sockfd), &unix),
-        _ => panic!("unrecognized SocketAddrAny kind"),
-    }) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn connect(
-    sockfd: c_int,
-    addr: *const SocketAddrStorage,
-    len: libc::socklen_t,
-) -> c_int {
-    // FIXME(#95) layout of addr doesn't match signature on i686
-    // uncomment once it does:
-    // libc!(libc::connect(sockfd, checked_cast!(addr), len));
-    libc!(libc::connect(sockfd, addr.cast(), len));
-
-    let addr = match convert_res(SocketAddrAny::read(addr, len.try_into().unwrap())) {
-        Some(addr) => addr,
-        None => return -1,
-    };
-    match convert_res(match addr {
-        SocketAddrAny::V4(v4) => rustix::net::connect_v4(&BorrowedFd::borrow_raw(sockfd), &v4),
-        SocketAddrAny::V6(v6) => rustix::net::connect_v6(&BorrowedFd::borrow_raw(sockfd), &v6),
-        SocketAddrAny::Unix(unix) => {
-            rustix::net::connect_unix(&BorrowedFd::borrow_raw(sockfd), &unix)
-        }
-        _ => panic!("unrecognized SocketAddrAny kind"),
-    }) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn getpeername(
-    fd: c_int,
-    addr: *mut SocketAddrStorage,
-    len: *mut libc::socklen_t,
-) -> c_int {
-    // FIXME(#95) layout of addr doesn't match signature on i686
-    // uncomment once it does:
-    // libc!(libc::getpeername(fd, checked_cast!(addr), len));
-    libc!(libc::getpeername(fd, addr.cast(), len));
-
-    match convert_res(rustix::net::getpeername(&BorrowedFd::borrow_raw(fd))) {
-        Some(from) => {
-            if let Some(from) = from {
-                let encoded_len = from.write(addr);
-                *len = encoded_len.try_into().unwrap();
-            }
-            0
-        }
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn getsockname(
-    fd: c_int,
-    addr: *mut SocketAddrStorage,
-    len: *mut libc::socklen_t,
-) -> c_int {
-    // FIXME(#95) layout of addr doesn't match signature on i686
-    // uncomment once it does:
-    // libc!(libc::getsockname(fd, checked_cast!(addr), len));
-    libc!(libc::getsockname(fd, addr.cast(), len));
-
-    match convert_res(rustix::net::getsockname(&BorrowedFd::borrow_raw(fd))) {
-        Some(from) => {
-            let encoded_len = from.write(addr);
-            *len = encoded_len.try_into().unwrap();
-            0
-        }
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn getsockopt(
-    fd: c_int,
-    level: c_int,
-    optname: c_int,
-    optval: *mut c_void,
-    optlen: *mut libc::socklen_t,
-) -> c_int {
-    use core::time::Duration;
-    use rustix::net::sockopt::{self, Timeout};
-
-    unsafe fn write_bool(
-        value: rustix::io::Result<bool>,
-        optval: *mut c_void,
-        optlen: *mut libc::socklen_t,
-    ) -> rustix::io::Result<()> {
-        Ok(write(value? as c_uint, optval.cast::<c_uint>(), optlen))
-    }
-
-    unsafe fn write_u32(
-        value: rustix::io::Result<u32>,
-        optval: *mut c_void,
-        optlen: *mut libc::socklen_t,
-    ) -> rustix::io::Result<()> {
-        Ok(write(value?, optval.cast::<u32>(), optlen))
-    }
-
-    unsafe fn write_linger(
-        linger: rustix::io::Result<Option<Duration>>,
-        optval: *mut c_void,
-        optlen: *mut libc::socklen_t,
-    ) -> rustix::io::Result<()> {
-        let linger = linger?;
-        let linger = libc::linger {
-            l_onoff: linger.is_some() as c_int,
-            l_linger: linger.unwrap_or_default().as_secs() as c_int,
-        };
-        Ok(write(linger, optval.cast::<libc::linger>(), optlen))
-    }
-
-    unsafe fn write_timeval(
-        value: rustix::io::Result<Option<Duration>>,
-        optval: *mut c_void,
-        optlen: *mut libc::socklen_t,
-    ) -> rustix::io::Result<()> {
-        let timeval = match value? {
-            None => libc::timeval {
-                tv_sec: 0,
-                tv_usec: 0,
-            },
-            Some(duration) => libc::timeval {
-                tv_sec: duration
-                    .as_secs()
-                    .try_into()
-                    .map_err(|_| rustix::io::Error::OVERFLOW)?,
-                tv_usec: duration.subsec_micros() as _,
-            },
-        };
-        Ok(write(timeval, optval.cast::<libc::timeval>(), optlen))
-    }
-
-    unsafe fn write<T>(value: T, optval: *mut T, optlen: *mut libc::socklen_t) {
-        *optlen = size_of::<T>().try_into().unwrap();
-        optval.write(value)
-    }
-
-    libc!(libc::getsockopt(fd, level, optname, optval, optlen));
-
-    let fd = &BorrowedFd::borrow_raw(fd);
-    let result = match level {
-        libc::SOL_SOCKET => match optname {
-            libc::SO_BROADCAST => write_bool(sockopt::get_socket_broadcast(fd), optval, optlen),
-            libc::SO_LINGER => write_linger(sockopt::get_socket_linger(fd), optval, optlen),
-            libc::SO_PASSCRED => write_bool(sockopt::get_socket_passcred(fd), optval, optlen),
-            libc::SO_SNDTIMEO => write_timeval(
-                sockopt::get_socket_timeout(fd, Timeout::Send),
-                optval,
-                optlen,
-            ),
-            libc::SO_RCVTIMEO => write_timeval(
-                sockopt::get_socket_timeout(fd, Timeout::Recv),
-                optval,
-                optlen,
-            ),
-            _ => unimplemented!("unimplemented getsockopt SOL_SOCKET optname {:?}", optname),
-        },
-        libc::IPPROTO_IP => match optname {
-            libc::IP_TTL => write_u32(sockopt::get_ip_ttl(fd), optval, optlen),
-            libc::IP_MULTICAST_LOOP => {
-                write_bool(sockopt::get_ip_multicast_loop(fd), optval, optlen)
-            }
-            libc::IP_MULTICAST_TTL => write_u32(sockopt::get_ip_multicast_ttl(fd), optval, optlen),
-            _ => unimplemented!("unimplemented getsockopt IPPROTO_IP optname {:?}", optname),
-        },
-        libc::IPPROTO_IPV6 => match optname {
-            libc::IPV6_MULTICAST_LOOP => {
-                write_bool(sockopt::get_ipv6_multicast_loop(fd), optval, optlen)
-            }
-            libc::IPV6_V6ONLY => write_bool(sockopt::get_ipv6_v6only(fd), optval, optlen),
-            _ => unimplemented!(
-                "unimplemented getsockopt IPPROTO_IPV6 optname {:?}",
-                optname
-            ),
-        },
-        libc::IPPROTO_TCP => match optname {
-            libc::TCP_NODELAY => write_bool(sockopt::get_tcp_nodelay(fd), optval, optlen),
-            _ => unimplemented!("unimplemented getsockopt IPPROTO_TCP optname {:?}", optname),
-        },
-        _ => unimplemented!(
-            "unimplemented getsockopt level {:?} optname {:?}",
-            level,
-            optname
-        ),
-    };
-    match convert_res(result) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn setsockopt(
-    fd: c_int,
-    level: c_int,
-    optname: c_int,
-    optval: *const c_void,
-    optlen: libc::socklen_t,
-) -> c_int {
-    use core::time::Duration;
-    use rustix::net::sockopt::{self, Timeout};
-
-    unsafe fn read_bool(optval: *const c_void, optlen: libc::socklen_t) -> bool {
-        read(optval.cast::<c_int>(), optlen) != 0
-    }
-
-    unsafe fn read_u32(optval: *const c_void, optlen: libc::socklen_t) -> u32 {
-        read(optval.cast::<u32>(), optlen)
-    }
-
-    unsafe fn read_linger(optval: *const c_void, optlen: libc::socklen_t) -> Option<Duration> {
-        let linger = read(optval.cast::<libc::linger>(), optlen);
-        (linger.l_onoff != 0).then(|| Duration::from_secs(linger.l_linger as u64))
-    }
-
-    unsafe fn read_timeval(optval: *const c_void, optlen: libc::socklen_t) -> Option<Duration> {
-        let timeval = read(optval.cast::<libc::timeval>(), optlen);
-        if timeval.tv_sec == 0 && timeval.tv_usec == 0 {
-            None
-        } else {
-            Some(
-                Duration::from_secs(timeval.tv_sec.try_into().unwrap())
-                    + Duration::from_micros(timeval.tv_usec as _),
-            )
-        }
-    }
-
-    unsafe fn read_ip_multiaddr(optval: *const c_void, optlen: libc::socklen_t) -> Ipv4Addr {
-        Ipv4Addr::from(
-            read(optval.cast::<libc::ip_mreq>(), optlen)
-                .imr_multiaddr
-                .s_addr,
-        )
-    }
-
-    unsafe fn read_ip_interface(optval: *const c_void, optlen: libc::socklen_t) -> Ipv4Addr {
-        Ipv4Addr::from(
-            read(optval.cast::<libc::ip_mreq>(), optlen)
-                .imr_interface
-                .s_addr,
-        )
-    }
-
-    unsafe fn read_ipv6_multiaddr(optval: *const c_void, optlen: libc::socklen_t) -> Ipv6Addr {
-        Ipv6Addr::from(
-            read(optval.cast::<libc::ipv6_mreq>(), optlen)
-                .ipv6mr_multiaddr
-                .s6_addr,
-        )
-    }
-
-    unsafe fn read_ipv6_interface(optval: *const c_void, optlen: libc::socklen_t) -> u32 {
-        read(optval.cast::<libc::ipv6_mreq>(), optlen).ipv6mr_interface
-    }
-
-    unsafe fn read<T>(optval: *const T, optlen: libc::socklen_t) -> T {
-        assert_eq!(optlen, size_of::<T>().try_into().unwrap());
-        optval.read()
-    }
-
-    libc!(libc::setsockopt(fd, level, optname, optval, optlen));
-
-    let fd = &BorrowedFd::borrow_raw(fd);
-    let result = match level {
-        libc::SOL_SOCKET => match optname {
-            libc::SO_REUSEADDR => sockopt::set_socket_reuseaddr(fd, read_bool(optval, optlen)),
-            libc::SO_BROADCAST => sockopt::set_socket_broadcast(fd, read_bool(optval, optlen)),
-            libc::SO_LINGER => sockopt::set_socket_linger(fd, read_linger(optval, optlen)),
-            libc::SO_PASSCRED => sockopt::set_socket_passcred(fd, read_bool(optval, optlen)),
-            libc::SO_SNDTIMEO => {
-                sockopt::set_socket_timeout(fd, Timeout::Send, read_timeval(optval, optlen))
-            }
-            libc::SO_RCVTIMEO => {
-                sockopt::set_socket_timeout(fd, Timeout::Recv, read_timeval(optval, optlen))
-            }
-            _ => unimplemented!("unimplemented setsockopt SOL_SOCKET optname {:?}", optname),
-        },
-        libc::IPPROTO_IP => match optname {
-            libc::IP_TTL => sockopt::set_ip_ttl(fd, read_u32(optval, optlen)),
-            libc::IP_MULTICAST_LOOP => {
-                sockopt::set_ip_multicast_loop(fd, read_bool(optval, optlen))
-            }
-            libc::IP_MULTICAST_TTL => sockopt::set_ip_multicast_ttl(fd, read_u32(optval, optlen)),
-            libc::IP_ADD_MEMBERSHIP => sockopt::set_ip_add_membership(
-                fd,
-                &read_ip_multiaddr(optval, optlen),
-                &read_ip_interface(optval, optlen),
-            ),
-            libc::IP_DROP_MEMBERSHIP => sockopt::set_ip_add_membership(
-                fd,
-                &read_ip_multiaddr(optval, optlen),
-                &read_ip_interface(optval, optlen),
-            ),
-            _ => unimplemented!("unimplemented setsockopt IPPROTO_IP optname {:?}", optname),
-        },
-        libc::IPPROTO_IPV6 => match optname {
-            libc::IPV6_MULTICAST_LOOP => {
-                sockopt::set_ipv6_multicast_loop(fd, read_bool(optval, optlen))
-            }
-            libc::IPV6_ADD_MEMBERSHIP => sockopt::set_ipv6_add_membership(
-                fd,
-                &read_ipv6_multiaddr(optval, optlen),
-                read_ipv6_interface(optval, optlen),
-            ),
-            libc::IPV6_DROP_MEMBERSHIP => sockopt::set_ipv6_drop_membership(
-                fd,
-                &read_ipv6_multiaddr(optval, optlen),
-                read_ipv6_interface(optval, optlen),
-            ),
-            libc::IPV6_V6ONLY => sockopt::set_ipv6_v6only(fd, read_bool(optval, optlen)),
-            _ => unimplemented!(
-                "unimplemented setsockopt IPPROTO_IPV6 optname {:?}",
-                optname
-            ),
-        },
-        libc::IPPROTO_TCP => match optname {
-            libc::TCP_NODELAY => sockopt::set_tcp_nodelay(fd, read_bool(optval, optlen)),
-            _ => unimplemented!("unimplemented setsockopt IPPROTO_TCP optname {:?}", optname),
-        },
-        _ => unimplemented!(
-            "unimplemented setsockopt level {:?} optname {:?}",
-            level,
-            optname
-        ),
-    };
-    match convert_res(result) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[cfg(feature = "sync-resolve")]
-#[no_mangle]
-unsafe extern "C" fn getaddrinfo(
-    node: *const c_char,
-    service: *const c_char,
-    hints: *const libc::addrinfo,
-    res: *mut *mut libc::addrinfo,
-) -> c_int {
-    libc!(libc::getaddrinfo(
-        node,
-        service,
-        checked_cast!(hints),
-        checked_cast!(res)
-    ));
-
-    assert!(service.is_null(), "service lookups not supported yet");
-    assert!(!node.is_null(), "only name lookups are supported corrently");
-
-    if !hints.is_null() {
-        let hints = &*hints;
-        assert_eq!(hints.ai_flags, 0, "GAI flags hint not supported yet");
-        assert_eq!(hints.ai_family, 0, "GAI family hint not supported yet");
-        assert_eq!(
-            hints.ai_socktype,
-            SocketType::STREAM.as_raw() as _,
-            "only SOCK_STREAM supported currently"
-        );
-        assert_eq!(hints.ai_protocol, 0, "GAI protocl hint not supported yet");
-        assert_eq!(hints.ai_addrlen, 0, "GAI addrlen hint not supported yet");
-        assert!(hints.ai_addr.is_null(), "GAI addr hint not supported yet");
-        assert!(
-            hints.ai_canonname.is_null(),
-            "GAI canonname hint not supported yet"
-        );
-        assert!(hints.ai_next.is_null(), "GAI next hint not supported yet");
-    }
-
-    let host = match ZStr::from_ptr(node.cast()).to_str() {
-        Ok(host) => host,
-        Err(_) => {
-            set_errno(Errno(libc::EILSEQ));
-            return libc::EAI_SYSTEM;
-        }
-    };
-
-    let layout = alloc::alloc::Layout::new::<libc::addrinfo>();
-    let addr_layout = alloc::alloc::Layout::new::<SocketAddrStorage>();
-    let mut first: *mut libc::addrinfo = null_mut();
-    let mut prev: *mut libc::addrinfo = null_mut();
-    match resolve_host(host) {
-        Ok(addrs) => {
-            for addr in addrs {
-                let ptr = alloc::alloc::alloc(layout).cast::<libc::addrinfo>();
-                ptr.write(zeroed());
-                let info = &mut *ptr;
-                match addr {
-                    IpAddr::V4(v4) => {
-                        // TODO: Create and write to `SocketAddrV4Storage`?
-                        let storage = alloc::alloc::alloc(addr_layout).cast::<SocketAddrStorage>();
-                        let len = SocketAddrAny::V4(SocketAddrV4::new(v4, 0)).write(storage);
-                        info.ai_addr = storage.cast();
-                        info.ai_addrlen = len.try_into().unwrap();
-                    }
-                    IpAddr::V6(v6) => {
-                        // TODO: Create and write to `SocketAddrV6Storage`?
-                        let storage = alloc::alloc::alloc(addr_layout).cast::<SocketAddrStorage>();
-                        let len = SocketAddrAny::V6(SocketAddrV6::new(v6, 0, 0, 0)).write(storage);
-                        info.ai_addr = storage.cast();
-                        info.ai_addrlen = len.try_into().unwrap();
-                    }
-                }
-                if !prev.is_null() {
-                    (*prev).ai_next = ptr;
-                }
-                prev = ptr;
-                if first.is_null() {
-                    first = ptr;
-                }
-            }
-            *res = first;
-            0
-        }
-        Err(err) => {
-            if let Some(err) = rustix::io::Error::from_io_error(&err) {
-                set_errno(Errno(err.raw_os_error()));
-                libc::EAI_SYSTEM
-            } else {
-                // TODO: sync-resolve should return a custom error type
-                if err.to_string()
-                    == "failed to resolve host: server responded with error: server failure"
-                    || err.to_string()
-                        == "failed to resolve host: server responded with error: no such name"
-                {
-                    libc::EAI_NONAME
-                } else {
-                    panic!("unknown error: {}", err);
-                }
-            }
-        }
-    }
-}
-
-#[cfg(feature = "sync-resolve")]
-#[no_mangle]
-unsafe extern "C" fn freeaddrinfo(mut res: *mut libc::addrinfo) {
-    libc!(libc::freeaddrinfo(checked_cast!(res)));
-
-    let layout = alloc::alloc::Layout::new::<libc::addrinfo>();
-    let addr_layout = alloc::alloc::Layout::new::<SocketAddrStorage>();
-
-    while !res.is_null() {
-        let addr = (*res).ai_addr;
-        if !addr.is_null() {
-            alloc::alloc::dealloc(addr.cast(), addr_layout);
-        }
-        let old = res;
-        res = (*res).ai_next;
-        alloc::alloc::dealloc(old.cast(), layout);
-    }
-}
-
-#[cfg(feature = "sync-resolve")]
-#[no_mangle]
-unsafe extern "C" fn gai_strerror(errcode: c_int) -> *const c_char {
-    libc!(libc::gai_strerror(errcode));
-
-    match errcode {
-        libc::EAI_NONAME => &b"Name does not resolve\0"[..],
-        libc::EAI_SYSTEM => &b"System error\0"[..],
-        _ => panic!("unrecognized gai_strerror {:?}", errcode),
-    }
-    .as_ptr()
-    .cast()
-}
-
-#[cfg(not(target_os = "wasi"))]
-#[no_mangle]
-unsafe extern "C" fn gethostname(name: *mut c_char, len: usize) -> c_int {
-    let uname = rustix::process::uname();
-    let nodename = uname.nodename();
-    if nodename.to_bytes().len() + 1 > len {
-        set_errno(Errno(libc::ENAMETOOLONG));
-        return -1;
-    }
-    memcpy(
-        name.cast(),
-        nodename.to_bytes().as_ptr().cast(),
-        nodename.to_bytes().len(),
-    );
-    *name.add(nodename.to_bytes().len()) = 0;
-    0
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn listen(fd: c_int, backlog: c_int) -> c_int {
-    libc!(libc::listen(fd, backlog));
-
-    match convert_res(rustix::net::listen(&BorrowedFd::borrow_raw(fd), backlog)) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn recv(fd: c_int, ptr: *mut c_void, len: usize, flags: c_int) -> isize {
-    libc!(libc::recv(fd, ptr, len, flags));
-
-    let flags = RecvFlags::from_bits(flags as _).unwrap();
-    match convert_res(rustix::net::recv(
-        &BorrowedFd::borrow_raw(fd),
-        slice::from_raw_parts_mut(ptr.cast::<u8>(), len),
-        flags,
-    )) {
-        Some(nread) => nread as isize,
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn recvfrom(
-    fd: c_int,
-    buf: *mut c_void,
-    len: usize,
-    flags: c_int,
-    from: *mut SocketAddrStorage,
-    from_len: *mut libc::socklen_t,
-) -> isize {
-    // FIXME(#95) layout of from doesn't match signature on i686
-    // uncomment once it does:
-    // libc!(libc::recvfrom(fd, buf, len, flags, checked_cast!(from), from_len));
-    libc!(libc::recvfrom(fd, buf, len, flags, from.cast(), from_len));
-
-    let flags = RecvFlags::from_bits(flags as _).unwrap();
-    match convert_res(rustix::net::recvfrom(
-        &BorrowedFd::borrow_raw(fd),
-        slice::from_raw_parts_mut(buf.cast::<u8>(), len),
-        flags,
-    )) {
-        Some((nread, addr)) => {
-            if let Some(addr) = addr {
-                let encoded_len = addr.write(from);
-                *from_len = encoded_len.try_into().unwrap();
-            }
-            nread as isize
-        }
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn send(fd: c_int, buf: *const c_void, len: usize, flags: c_int) -> isize {
-    libc!(libc::send(fd, buf, len, flags));
-
-    let flags = SendFlags::from_bits(flags as _).unwrap();
-    match convert_res(rustix::net::send(
-        &BorrowedFd::borrow_raw(fd),
-        slice::from_raw_parts(buf.cast::<u8>(), len),
-        flags,
-    )) {
-        Some(nwritten) => nwritten as isize,
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn sendto(
-    fd: c_int,
-    buf: *const c_void,
-    len: usize,
-    flags: c_int,
-    to: *const SocketAddrStorage,
-    to_len: libc::socklen_t,
-) -> isize {
-    // FIXME(#95) layout of to doesn't match signature on i686
-    // uncomment once it does:
-    // libc!(libc::sendto(fd, buf, len, flags, checked_cast!(to), to_len));
-    libc!(libc::sendto(fd, buf, len, flags, to.cast(), to_len));
-
-    let flags = SendFlags::from_bits(flags as _).unwrap();
-    let addr = match convert_res(SocketAddrAny::read(to, to_len.try_into().unwrap())) {
-        Some(addr) => addr,
-        None => return -1,
-    };
-    match convert_res(match addr {
-        SocketAddrAny::V4(v4) => rustix::net::sendto_v4(
-            &BorrowedFd::borrow_raw(fd),
-            slice::from_raw_parts(buf.cast::<u8>(), len),
-            flags,
-            &v4,
-        ),
-        SocketAddrAny::V6(v6) => rustix::net::sendto_v6(
-            &BorrowedFd::borrow_raw(fd),
-            slice::from_raw_parts(buf.cast::<u8>(), len),
-            flags,
-            &v6,
-        ),
-        SocketAddrAny::Unix(unix) => rustix::net::sendto_unix(
-            &BorrowedFd::borrow_raw(fd),
-            slice::from_raw_parts(buf.cast::<u8>(), len),
-            flags,
-            &unix,
-        ),
-        _ => panic!("unrecognized SocketAddrAny kind"),
-    }) {
-        Some(nwritten) => nwritten as isize,
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn shutdown(fd: c_int, how: c_int) -> c_int {
-    libc!(libc::shutdown(fd, how));
-
-    let how = match how {
-        libc::SHUT_RD => Shutdown::Read,
-        libc::SHUT_WR => Shutdown::Write,
-        libc::SHUT_RDWR => Shutdown::ReadWrite,
-        _ => panic!("unrecognized shutdown kind {}", how),
-    };
-    match convert_res(rustix::net::shutdown(&BorrowedFd::borrow_raw(fd), how)) {
-        Some(()) => 0,
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn socket(domain: c_int, type_: c_int, protocol: c_int) -> c_int {
-    libc!(libc::socket(domain, type_, protocol));
-
-    let domain = AddressFamily::from_raw(domain as _);
-    let flags = SocketFlags::from_bits_truncate(type_ as _);
-    let type_ = SocketType::from_raw(type_ as u32 & !SocketFlags::all().bits());
-    let protocol = Protocol::from_raw(protocol as _);
-    match convert_res(rustix::net::socket_with(domain, type_, flags, protocol)) {
-        Some(fd) => fd.into_raw_fd(),
-        None => -1,
-    }
-}
-
-#[cfg(feature = "net")]
-#[no_mangle]
-unsafe extern "C" fn socketpair(
-    domain: c_int,
-    type_: c_int,
-    protocol: c_int,
-    sv: *mut [c_int; 2],
-) -> c_int {
-    libc!(libc::socketpair(
-        domain,
-        type_,
-        protocol,
-        (*sv).as_mut_ptr()
-    ));
-
-    let domain = AddressFamily::from_raw(domain as _);
-    let flags = SocketFlags::from_bits_truncate(type_ as _);
-    let type_ = SocketType::from_raw(type_ as u32 & !SocketFlags::all().bits());
-    let protocol = Protocol::from_raw(protocol as _);
-    match convert_res(rustix::net::socketpair(domain, type_, flags, protocol)) {
-        Some((fd0, fd1)) => {
-            (*sv) = [fd0.into_raw_fd(), fd1.into_raw_fd()];
-            0
-        }
-        None => -1,
-    }
-}
-
-#[cfg(feature = "sync-resolve")]
-#[no_mangle]
-unsafe extern "C" fn __res_init() -> c_int {
-    libc!(libc::res_init());
     0
 }
 
@@ -1107,26 +334,6 @@ unsafe extern "C" fn getrandom(buf: *mut c_void, buflen: usize, flags: u32) -> i
     }
 }
 
-// termios
-
-#[no_mangle]
-unsafe extern "C" fn tcgetattr() {
-    //libc!(libc::tcgetattr());
-    unimplemented!("tcgetattr")
-}
-
-#[no_mangle]
-unsafe extern "C" fn tcsetattr() {
-    //libc!(libc::tcsetattr());
-    unimplemented!("tcsetattr")
-}
-
-#[no_mangle]
-unsafe extern "C" fn cfmakeraw() {
-    //libc!(libc::cfmakeraw());
-    unimplemented!("cfmakeraw")
-}
-
 // process
 
 #[no_mangle]
@@ -1254,24 +461,6 @@ unsafe extern "C" fn dlerror() {
 unsafe extern "C" fn dlopen() {
     //libc!(libc::dlopen());
     unimplemented!("dlopen")
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn __tls_get_addr(p: &[usize; 2]) -> *mut c_void {
-    //libc!(libc::__tls_get_addr(p));
-    let [module, offset] = *p;
-    // Offset 0 is the generation field, and we don't support dynamic linking,
-    // so we should only sever see 1 here.
-    assert_eq!(module, 1);
-    origin::current_thread_tls_addr(offset)
-}
-
-#[cfg(target_arch = "x86")]
-#[no_mangle]
-unsafe extern "C" fn ___tls_get_addr() {
-    //libc!(libc::___tls_get_addr());
-    unimplemented!("___tls_get_addr")
 }
 
 #[no_mangle]
@@ -2413,768 +1602,6 @@ fn init_from_envp(envp: *mut *mut c_char) {
 
 #[no_mangle]
 static mut environ: *mut *mut c_char = null_mut();
-
-// Threads
-#[cfg(feature = "threads")]
-struct GetThreadId;
-
-#[cfg(feature = "threads")]
-unsafe impl parking_lot::lock_api::GetThreadId for GetThreadId {
-    const INIT: Self = Self;
-
-    fn nonzero_thread_id(&self) -> core::num::NonZeroUsize {
-        origin::current_thread_id()
-            .as_raw_nonzero()
-            .try_into()
-            .unwrap()
-    }
-}
-
-// In Linux, `pthread_t` is usually `unsigned long`, but we make it a pointer
-// type so that it preserves provenance.
-#[cfg(feature = "threads")]
-#[allow(non_camel_case_types)]
-type PthreadT = *mut c_void;
-#[cfg(feature = "threads")]
-libc_type!(PthreadT, pthread_t);
-
-#[cfg(feature = "threads")]
-#[allow(non_camel_case_types)]
-#[repr(C)]
-#[derive(Clone)]
-struct PthreadAttrT {
-    stack_addr: *mut c_void,
-    stack_size: usize,
-    guard_size: usize,
-    pad0: usize,
-    pad1: usize,
-    pad2: usize,
-    pad3: usize,
-    #[cfg(target_arch = "x86")]
-    pad4: usize,
-    #[cfg(target_arch = "x86")]
-    pad5: usize,
-}
-#[cfg(feature = "threads")]
-libc_type!(PthreadAttrT, pthread_attr_t);
-
-#[cfg(feature = "threads")]
-impl Default for PthreadAttrT {
-    fn default() -> Self {
-        Self {
-            stack_addr: null_mut(),
-            stack_size: origin::default_stack_size(),
-            guard_size: origin::default_guard_size(),
-            pad0: 0,
-            pad1: 0,
-            pad2: 0,
-            pad3: 0,
-            #[cfg(target_arch = "x86")]
-            pad4: 0,
-            #[cfg(target_arch = "x86")]
-            pad5: 0,
-        }
-    }
-}
-
-#[cfg(feature = "threads")]
-#[allow(non_camel_case_types)]
-#[repr(C)]
-union pthread_mutex_u {
-    // Use a custom `RawMutex` for non-reentrant pthread mutex for now, because
-    // we use `dlmalloc` or `wee_alloc` for global allocations, which use
-    // mutexes, and `parking_lot`'s `RawMutex` performs global allocations.
-    normal: RawMutex,
-    reentrant: parking_lot::lock_api::RawReentrantMutex<parking_lot::RawMutex, GetThreadId>,
-}
-
-#[cfg(feature = "threads")]
-#[allow(non_camel_case_types)]
-#[repr(C)]
-struct PthreadMutexT {
-    kind: AtomicU32,
-    u: pthread_mutex_u,
-    pad0: usize,
-    #[cfg(target_arch = "x86")]
-    pad1: usize,
-}
-#[cfg(feature = "threads")]
-libc_type!(PthreadMutexT, pthread_mutex_t);
-
-#[cfg(feature = "threads")]
-#[allow(non_camel_case_types)]
-#[repr(C)]
-struct PthreadRwlockT {
-    lock: parking_lot::RawRwLock,
-    exclusive: AtomicBool,
-    pad0: usize,
-    pad1: usize,
-    pad2: usize,
-    pad3: usize,
-    pad4: usize,
-    #[cfg(target_arch = "x86")]
-    pad5: usize,
-}
-#[cfg(feature = "threads")]
-libc_type!(PthreadRwlockT, pthread_rwlock_t);
-
-#[cfg(feature = "threads")]
-#[allow(non_camel_case_types)]
-#[repr(C)]
-struct PthreadMutexattrT {
-    kind: AtomicU32,
-}
-#[cfg(feature = "threads")]
-libc_type!(PthreadMutexattrT, pthread_mutexattr_t);
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_self() -> PthreadT {
-    libc!(ptr::from_exposed_addr_mut(libc::pthread_self() as _));
-    origin::current_thread().cast()
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_getattr_np(thread: PthreadT, attr: *mut PthreadAttrT) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_getattr_np(thread, checked_cast!(attr)));
-    let (stack_addr, stack_size, guard_size) = origin::thread_stack(thread.cast());
-    ptr::write(
-        attr,
-        PthreadAttrT {
-            stack_addr,
-            stack_size,
-            guard_size,
-            pad0: 0,
-            pad1: 0,
-            pad2: 0,
-            pad3: 0,
-            #[cfg(target_arch = "x86")]
-            pad4: 0,
-            #[cfg(target_arch = "x86")]
-            pad5: 0,
-        },
-    );
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_attr_init(attr: *mut PthreadAttrT) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_attr_init(checked_cast!(attr)));
-    ptr::write(attr, PthreadAttrT::default());
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_attr_destroy(attr: *mut PthreadAttrT) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_attr_destroy(checked_cast!(attr)));
-    ptr::drop_in_place(attr);
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_attr_getstack(
-    attr: *const PthreadAttrT,
-    stackaddr: *mut *mut c_void,
-    stacksize: *mut usize,
-) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_attr_getstack(checked_cast!(attr), stackaddr, stacksize));
-    *stackaddr = (*attr).stack_addr;
-    *stacksize = (*attr).stack_size;
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_getspecific() -> *const c_void {
-    //libc!(libc::pthread_getspecific());
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_getspecific\n",
-    )
-    .ok();
-    null()
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_key_create() -> c_int {
-    //libc!(libc::pthread_key_create());
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_key_create\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_key_delete() -> c_int {
-    //libc!(libc::pthread_key_delete());
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_key_delete\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_mutexattr_destroy(attr: *mut PthreadMutexattrT) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_mutexattr_destroy(checked_cast!(attr)));
-    ptr::drop_in_place(attr);
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_mutexattr_init(attr: *mut PthreadMutexattrT) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_mutexattr_init(checked_cast!(attr)));
-    ptr::write(
-        attr,
-        PthreadMutexattrT {
-            kind: AtomicU32::new(libc::PTHREAD_MUTEX_DEFAULT as u32),
-        },
-    );
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_mutexattr_settype(attr: *mut PthreadMutexattrT, kind: c_int) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_mutexattr_settype(checked_cast!(attr), kind));
-    (*attr).kind = AtomicU32::new(kind as u32);
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_mutex_destroy(mutex: *mut PthreadMutexT) -> c_int {
-    // FIXME(#95) layout of mutex doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_mutex_destroy(checked_cast!(mutex)));
-    match (*mutex).kind.load(SeqCst) as i32 {
-        libc::PTHREAD_MUTEX_NORMAL => ptr::drop_in_place(&mut (*mutex).u.normal),
-        libc::PTHREAD_MUTEX_RECURSIVE => ptr::drop_in_place(&mut (*mutex).u.reentrant),
-        libc::PTHREAD_MUTEX_ERRORCHECK => unimplemented!("PTHREAD_MUTEX_ERRORCHECK"),
-        other => unimplemented!("unsupported pthread mutex kind {}", other),
-    }
-    (*mutex).kind.store(!0, SeqCst);
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_mutex_init(
-    mutex: *mut PthreadMutexT,
-    mutexattr: *const PthreadMutexattrT,
-) -> c_int {
-    // FIXME(#95) layout of mutex and attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_mutex_init(checked_cast!(mutex), checked_cast!(mutexattr)));
-    let kind = (*mutexattr).kind.load(SeqCst);
-    match kind as i32 {
-        libc::PTHREAD_MUTEX_NORMAL => ptr::write(&mut (*mutex).u.normal, RawMutex::INIT),
-        libc::PTHREAD_MUTEX_RECURSIVE => ptr::write(
-            &mut (*mutex).u.reentrant,
-            parking_lot::lock_api::RawReentrantMutex::<parking_lot::RawMutex, GetThreadId>::INIT,
-        ),
-        libc::PTHREAD_MUTEX_ERRORCHECK => unimplemented!("PTHREAD_MUTEX_ERRORCHECK"),
-        other => unimplemented!("unsupported pthread mutex kind {}", other),
-    }
-    (*mutex).kind.store(kind, SeqCst);
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_mutex_lock(mutex: *mut PthreadMutexT) -> c_int {
-    // FIXME(#95) layout of mutex doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_mutex_lock(checked_cast!(mutex)));
-    match (*mutex).kind.load(SeqCst) as i32 {
-        libc::PTHREAD_MUTEX_NORMAL => (*mutex).u.normal.lock(),
-        libc::PTHREAD_MUTEX_RECURSIVE => (*mutex).u.reentrant.lock(),
-        libc::PTHREAD_MUTEX_ERRORCHECK => unimplemented!("PTHREAD_MUTEX_ERRORCHECK"),
-        other => unimplemented!("unsupported pthread mutex kind {}", other),
-    }
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_mutex_trylock(mutex: *mut PthreadMutexT) -> c_int {
-    // FIXME(#95) layout of mutex doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_mutex_trylock(checked_cast!(mutex)));
-    if match (*mutex).kind.load(SeqCst) as i32 {
-        libc::PTHREAD_MUTEX_NORMAL => (*mutex).u.normal.try_lock(),
-        libc::PTHREAD_MUTEX_RECURSIVE => (*mutex).u.reentrant.try_lock(),
-        libc::PTHREAD_MUTEX_ERRORCHECK => unimplemented!("PTHREAD_MUTEX_ERRORCHECK"),
-        other => unimplemented!("unsupported pthread mutex kind {}", other),
-    } {
-        0
-    } else {
-        rustix::io::Error::BUSY.raw_os_error()
-    }
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_mutex_unlock(mutex: *mut PthreadMutexT) -> c_int {
-    // FIXME(#95) layout of mutex doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_mutex_unlock(checked_cast!(mutex)));
-    match (*mutex).kind.load(SeqCst) as i32 {
-        libc::PTHREAD_MUTEX_NORMAL => (*mutex).u.normal.unlock(),
-        libc::PTHREAD_MUTEX_RECURSIVE => (*mutex).u.reentrant.unlock(),
-        libc::PTHREAD_MUTEX_ERRORCHECK => unimplemented!("PTHREAD_MUTEX_ERRORCHECK"),
-        other => unimplemented!("unsupported pthread mutex kind {}", other),
-    }
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_tryrdlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_tryrdlock(checked_cast!(rwlock)));
-    let result = (*rwlock).lock.try_lock_shared();
-    if result {
-        (*rwlock).exclusive.store(false, SeqCst);
-        0
-    } else {
-        rustix::io::Error::BUSY.raw_os_error()
-    }
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_trywrlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_trywrlock(checked_cast!(rwlock)));
-    let result = (*rwlock).lock.try_lock_exclusive();
-    if result {
-        (*rwlock).exclusive.store(true, SeqCst);
-        0
-    } else {
-        rustix::io::Error::BUSY.raw_os_error()
-    }
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_rdlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_rdlock(checked_cast!(rwlock)));
-    (*rwlock).lock.lock_shared();
-    (*rwlock).exclusive.store(false, SeqCst);
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_unlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_unlock(checked_cast!(rwlock)));
-    if (*rwlock).exclusive.load(SeqCst) {
-        (*rwlock).lock.unlock_exclusive();
-    } else {
-        (*rwlock).lock.unlock_shared();
-    }
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_setspecific() -> c_int {
-    //libc!(libc::pthread_setspecific());
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_getspecific\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_attr_getguardsize(
-    attr: *const PthreadAttrT,
-    guardsize: *mut usize,
-) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_attr_getguardsize(checked_cast!(attr), guardsize));
-    *guardsize = (*attr).guard_size;
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_attr_setguardsize(attr: *mut PthreadAttrT, guardsize: usize) -> c_int {
-    // TODO: libc!(libc::pthread_attr_setguardsize(attr, guardsize));
-    (*attr).guard_size = guardsize;
-    0
-}
-
-#[cfg(feature = "threads")]
-const SIZEOF_PTHREAD_COND_T: usize = 48;
-
-#[cfg(feature = "threads")]
-#[cfg_attr(target_arch = "x86", repr(C, align(4)))]
-#[cfg_attr(not(target_arch = "x86"), repr(C, align(8)))]
-struct PthreadCondT {
-    inner: parking_lot::Condvar,
-    pad: [u8; SIZEOF_PTHREAD_COND_T - core::mem::size_of::<parking_lot::Condvar>()],
-}
-
-#[cfg(feature = "threads")]
-libc_type!(PthreadCondT, pthread_cond_t);
-
-#[cfg(feature = "threads")]
-#[cfg(any(
-    target_arch = "x86",
-    target_arch = "x86_64",
-    target_arch = "arm",
-    all(target_arch = "aarch64", target_pointer_width = "32"),
-    target_arch = "riscv64",
-))]
-const SIZEOF_PTHREAD_CONDATTR_T: usize = 4;
-
-#[cfg(all(target_arch = "aarch64", target_pointer_width = "64"))]
-const SIZEOF_PTHREAD_CONDATTR_T: usize = 8;
-
-#[cfg(feature = "threads")]
-#[repr(C, align(4))]
-struct PthreadCondattrT {
-    pad: [u8; SIZEOF_PTHREAD_CONDATTR_T],
-}
-
-#[cfg(feature = "threads")]
-libc_type!(PthreadCondattrT, pthread_condattr_t);
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_condattr_destroy(attr: *mut PthreadCondattrT) -> c_int {
-    libc!(libc::pthread_condattr_destroy(checked_cast!(attr)));
-    let _ = attr;
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_condattr_destroy\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_condattr_init(attr: *mut PthreadCondattrT) -> c_int {
-    libc!(libc::pthread_condattr_init(checked_cast!(attr)));
-    let _ = attr;
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_condattr_init\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_condattr_setclock(
-    attr: *mut PthreadCondattrT,
-    clock_id: c_int,
-) -> c_int {
-    libc!(libc::pthread_condattr_setclock(
-        checked_cast!(attr),
-        clock_id
-    ));
-    let _ = (attr, clock_id);
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_condattr_setclock\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_cond_broadcast(cond: *mut PthreadCondT) -> c_int {
-    libc!(libc::pthread_cond_broadcast(checked_cast!(cond)));
-    let _ = cond;
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_cond_broadcast\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_cond_destroy(cond: *mut PthreadCondT) -> c_int {
-    libc!(libc::pthread_cond_destroy(checked_cast!(cond)));
-    let _ = cond;
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_cond_destroy\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_cond_init(
-    cond: *mut PthreadCondT,
-    attr: *const PthreadCondattrT,
-) -> c_int {
-    libc!(libc::pthread_cond_init(
-        checked_cast!(cond),
-        checked_cast!(attr)
-    ));
-    let _ = (cond, attr);
-    rustix::io::write(&rustix::io::stderr(), b"unimplemented: pthread_cond_init\n").ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_cond_signal(cond: *mut PthreadCondT) -> c_int {
-    libc!(libc::pthread_cond_signal(checked_cast!(cond)));
-    let _ = cond;
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_cond_signal\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_cond_wait(cond: *mut PthreadCondT, lock: *mut PthreadMutexT) -> c_int {
-    // FIXME(#95) layout of lock doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_cond_wait(checked_cast!(cond), checked_cast!(lock));
-    let _ = (cond, lock);
-    rustix::io::write(&rustix::io::stderr(), b"unimplemented: pthread_cond_wait\n").ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_cond_timedwait(
-    cond: *mut PthreadCondT,
-    lock: *mut PthreadMutexT,
-    abstime: *const libc::timespec,
-) -> c_int {
-    // FIXME(#95) layout of mutex doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_cond_timedwait(checked_cast!(cond), checked_cast!(lock), abstime));
-    let _ = (cond, lock, abstime);
-    rustix::io::write(
-        &rustix::io::stderr(),
-        b"unimplemented: pthread_cond_timedwait\n",
-    )
-    .ok();
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_destroy(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_destroy(checked_cast!(rwlock)));
-    ptr::drop_in_place(rwlock);
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_rwlock_wrlock(rwlock: *mut PthreadRwlockT) -> c_int {
-    libc!(libc::pthread_rwlock_wrlock(checked_cast!(rwlock)));
-    (*rwlock).lock.lock_exclusive();
-    (*rwlock).exclusive.store(true, SeqCst);
-    0
-}
-
-// TODO: signals, create thread-local storage, arrange for thread-local storage
-// destructors to run, free the `mmap`.
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_create(
-    pthread: *mut PthreadT,
-    attr: *const PthreadAttrT,
-    fn_: extern "C" fn(*mut c_void) -> *mut c_void,
-    arg: *mut c_void,
-) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_create(pthread, checked_cast!(attr), fn_, arg));
-    let PthreadAttrT {
-        stack_addr,
-        stack_size,
-        guard_size,
-        pad0: _,
-        pad1: _,
-        pad2: _,
-        pad3: _,
-        #[cfg(target_arch = "x86")]
-            pad4: _,
-        #[cfg(target_arch = "x86")]
-            pad5: _,
-    } = if attr.is_null() {
-        PthreadAttrT::default()
-    } else {
-        (*attr).clone()
-    };
-    assert!(stack_addr.is_null());
-
-    let thread = match origin::create_thread(
-        Box::new(move || {
-            fn_(arg);
-            Some(Box::new(arg))
-        }),
-        stack_size,
-        guard_size,
-    ) {
-        Ok(thread) => thread,
-        Err(e) => return e.raw_os_error(),
-    };
-
-    pthread.write(thread.cast());
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_detach(pthread: PthreadT) -> c_int {
-    libc!(libc::pthread_detach(pthread.expose_addr() as _));
-    origin::detach_thread(pthread.cast());
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_join(pthread: PthreadT, retval: *mut *mut c_void) -> c_int {
-    libc!(libc::pthread_join(pthread.expose_addr() as _, retval));
-    assert!(
-        retval.is_null(),
-        "pthread_join return values not supported yet"
-    );
-
-    origin::join_thread(pthread.cast());
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_sigmask() -> c_int {
-    //libc!(libc::pthread_sigmask());
-    // not yet implemented, what is needed fo `std::Command` to work.
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_attr_setstacksize(attr: *mut PthreadAttrT, stacksize: usize) -> c_int {
-    // FIXME(#95) layout of attr doesn't match signature on aarch64
-    // uncomment once it does:
-    // libc!(libc::pthread_attr_setstacksize(checked_cast!(attr), stacksize));
-    (*attr).stack_size = stacksize;
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_cancel() -> c_int {
-    //libc!(libc::pthread_cancel());
-    unimplemented!("pthread_cancel")
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_exit() -> c_int {
-    //libc!(libc::pthread_exit());
-    // As with `pthread_cancel`, `pthread_exit` may be tricky to implement.
-    unimplemented!("pthread_exit")
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_cleanup_push() -> c_int {
-    //libc!(libc::pthread_cleanup_push());
-    unimplemented!("pthread_cleanup_push")
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_cleanup_pop() -> c_int {
-    //libc!(libc::pthread_cleanup_pop());
-    unimplemented!("pthread_cleanup_pop")
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_setcancelstate() -> c_int {
-    //libc!(libc::pthread_setcancelstate());
-    unimplemented!("pthread_setcancelstate")
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_setcanceltype() -> c_int {
-    //libc!(libc::pthread_setcanceltype());
-    unimplemented!("pthread_setcanceltype")
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_testcancel() -> c_int {
-    //libc!(libc::pthread_testcancel());
-    unimplemented!("pthread_testcancel")
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn pthread_atfork(
-    prepare: Option<unsafe extern "C" fn()>,
-    parent: Option<unsafe extern "C" fn()>,
-    child: Option<unsafe extern "C" fn()>,
-) -> c_int {
-    libc!(libc::pthread_atfork(prepare, parent, child));
-    at_fork::at_fork(prepare, parent, child);
-    0
-}
-
-#[cfg(feature = "threads")]
-#[no_mangle]
-unsafe extern "C" fn __cxa_thread_atexit_impl(
-    func: unsafe extern "C" fn(*mut c_void),
-    obj: *mut c_void,
-    _dso_symbol: *mut c_void,
-) -> c_int {
-    // TODO: libc!(libc::__cxa_thread_atexit_impl(func, obj, _dso_symbol));
-    origin::at_thread_exit_raw(func, obj);
-    0
-}
 
 // exit
